# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pyram Sterling Bank API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List my transactions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: type
          in: query
          schema:
            type: string
            enum: [deposit, withdrawal, transfer, loan_disbursement, loan_repayment, partner_payment]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /transactions/deposit:
    post:
      operationId: deposit
      tags: [transactions]
      summary: Deposit funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/withdraw:
    post:
      operationId: withdraw
      tags: [transactions]
      summary: Withdraw funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawInput"
      responses:
        "201":
          description: Withdrawal recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/transfer:
    post:
      operationId: transfer
      tags: [transactions]
      summary: Transfer money to another account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferInput"
      responses:
        "201":
          description: Transfer completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /admin/transactions:
    get:
      operationId: adminListTransactions
      tags: [admin]
      summary: Admin — list all transactions globally
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: userId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed]
        - name: type
          in: query
          schema:
            type: string
            enum: [deposit, withdrawal, transfer, loan_disbursement, loan_repayment, partner_payment]
      responses:
        "200":
          description: All transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, userId, type, amount, description, status, createdAt]
      properties:
        id:               { type: integer }
        userId:           { type: integer }
        type:             { type: string, enum: [deposit, withdrawal, transfer, loan_disbursement, loan_repayment, partner_payment] }
        amount:           { type: number }
        description:      { type: string }
        status:           { type: string, enum: [completed, pending, failed] }
        recipientAccount: { type: ["string", "null"] }
        recipientName:    { type: ["string", "null"] }
        senderAccount:    { type: ["string", "null"] }
        createdAt:        { type: string }
...
    DepositInput:
      type: object
      required: [amount]
      properties:
        amount:      { type: number, minimum: 0.01 }
        description: { type: string }
    WithdrawInput:
      type: object
      required: [amount]
      properties:
        amount:      { type: number, minimum: 0.01 }
        description: { type: string }
    TransferInput:
      type: object
      required: [amount, recipientAccount]
      properties:
        amount:           { type: number, minimum: 0.01 }
        recipientAccount: { type: string }
        description:      { type: string }
...